// src/app/esp-activity-module/models/tactic.ts
import { User } from "./user";
import { AllowedAction } from "./allowed-action";
import { Tag } from "./tag";
import { Activity } from "./activity";
import { Category } from "./category";
export class Tactic {
  constructor() {}
  id?: number;
  title?: string;
  definitionVal?: number;
  conditionVal?: number;
  definition?: string;
  condition?: string;
  startDate?: Date;
  dueDate?: Date;
  status?: number;
  weight?: number;
  organizationId?: number;
  //organization?: Organization;
  teamId?: number;
  // team?: Team;
  indicatorId?: Number;
  indicator?: Activity;
  categoryId?: number;
  category?: Category;
  notes?: string;
  createdBy_userId?: number;
  createdBy?: User;
  assignedBy_UserId?: number;
  assignedBy?: User;
  parentName?: string;
  unit?: string;
  creationDate?: Date;
  score?: number;
  scoreCached?: number;
  indicatorCount?: number;
  relationCount?: number;
  allowedActions?: AllowedAction[];
  labelGroupId?: number;
  owner_UserId?: number;
  userOwner?: User;
  targetValue?: number;
  actualValue?: number;
  fieldCount?: number;
  signatureCount?: number;
  hasLoggedInUserSign?: boolean;
  isSigned?: boolean;
  isCascaded?: boolean;
  attachmentCount?: number;
  ratedIndicatorCount?: number;
  commentCount?: number;
  isLocked?: boolean;
  tags?: Tag[];
  hasCascadedTactic?: boolean;
  IndicatorParentName?: string;
  templateId?: number;
}